DB-11955 Add scan cost estimation to pluggable cost estimation model. (#5427)
* DB-11955 Add scan cost estimation to pluggable cost estimation model.

* DB-11955 minor fix.

* DB-11955 refactoring.

* DB-11955 refactoring.
