+++ db-engine/src/main/java/com/splicemachine/db/iapi/sql/compile/Optimizer.java
@@ -32,8 +32,7 @@
 package com.splicemachine.db.iapi.sql.compile;
 
 import com.splicemachine.db.iapi.error.StandardException;
-import com.splicemachine.db.iapi.sql.compile.costing.JoinCostEstimationModel;
-import com.splicemachine.db.iapi.sql.dictionary.ConglomerateDescriptor;
+import com.splicemachine.db.iapi.sql.compile.costing.CostModel;
 import com.splicemachine.db.iapi.sql.dictionary.DataDictionary;
 import com.splicemachine.db.iapi.sql.dictionary.TableDescriptor;
 import com.splicemachine.db.iapi.store.access.AggregateCostController;
@@ -425,5 +424,5 @@ public interface Optimizer{
 
     default boolean isMemPlatform() { return false; };
 
-    JoinCostEstimationModel getJoinCostEstimationModel();
+    CostModel getCostModel();
 }
