+++ db-engine/src/main/java/com/splicemachine/db/iapi/sql/conn/SessionProperties.java
@@ -32,7 +32,7 @@ package com.splicemachine.db.iapi.sql.conn;
 
 import com.splicemachine.db.iapi.error.StandardException;
 import com.splicemachine.db.iapi.reference.SQLState;
-import com.splicemachine.db.iapi.sql.compile.costing.JoinCostEstimationModelRegistry;
+import com.splicemachine.db.iapi.sql.compile.costing.CostModelRegistry;
 import com.splicemachine.utils.Pair;
 
 import java.sql.SQLException;
@@ -62,7 +62,7 @@ public interface SessionProperties {
         ALWAYSALLOWINDEXPREFIXITERATION(17),
         OLAPALWAYSPENALIZENLJ(18),
         FAVORINDEXPREFIXITERATION(19),
-        JOINCOSTESTIMATIONMODEL(20);
+        COSTMODEL(20);
 
         public static final int COUNT = PROPERTYNAME.values().length;
 
@@ -96,7 +96,7 @@ public interface SessionProperties {
             property = SessionProperties.PROPERTYNAME.valueOf(propertyNameString);
         } catch (IllegalArgumentException e) {
             throw StandardException.newException(SQLState.LANG_INVALID_SESSION_PROPERTY,propertyNameString,
-                "useOLAP, useSpark (deprecated), defaultSelectivityFactor, skipStats, olapQueue, recursiveQueryIterationLimit, tableLimitForExhaustiveSearch, minPlanTimeout, currentFunctionPath, disablePredsForIndexOrPkAccessPath, alwaysAllowIndexPrefixIteration, olapAlwaysPenalizeNLJ, favorIndexPrefixIteration,joincostEstimationModel");
+                "useOLAP, useSpark (deprecated), defaultSelectivityFactor, skipStats, olapQueue, recursiveQueryIterationLimit, tableLimitForExhaustiveSearch, minPlanTimeout, currentFunctionPath, disablePredsForIndexOrPkAccessPath, alwaysAllowIndexPrefixIteration, olapAlwaysPenalizeNLJ, favorIndexPrefixIteration, costModel");
         }
 
         String valString = pair.getSecond();
@@ -156,9 +156,9 @@ public interface SessionProperties {
                     throw StandardException.newException(SQLState.LANG_INVALID_SESSION_PROPERTY_VALUE, valString, "value should be a positive long");
                 }
                 break;
-            case JOINCOSTESTIMATIONMODEL:
-                if(!JoinCostEstimationModelRegistry.exists(valString)) {
-                    throw StandardException.newException(SQLState.LANG_INVALID_SESSION_PROPERTY_VALUE, valString, String.format("value should be one of %s", JoinCostEstimationModelRegistry.modelNames()));
+            case COSTMODEL:
+                if(!CostModelRegistry.exists(valString)) {
+                    throw StandardException.newException(SQLState.LANG_INVALID_SESSION_PROPERTY_VALUE, valString, String.format("value should be one of %s", CostModelRegistry.modelNames()));
                 }
             default:
                 break;
