+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/ResultSetNode.java
@@ -40,7 +40,6 @@ import com.splicemachine.db.iapi.services.sanity.SanityManager;
 import com.splicemachine.db.iapi.sql.ResultColumnDescriptor;
 import com.splicemachine.db.iapi.sql.ResultDescription;
 import com.splicemachine.db.iapi.sql.compile.*;
-import com.splicemachine.db.iapi.sql.compile.costing.JoinCostEstimationModelRegistry;
 import com.splicemachine.db.iapi.sql.conn.LanguageConnectionContext;
 import com.splicemachine.db.iapi.sql.dictionary.ColumnDescriptor;
 import com.splicemachine.db.iapi.sql.dictionary.DataDictionary;
@@ -1130,7 +1129,7 @@ public abstract class ResultSetNode extends QueryTreeNode{
                                                     requiredRowOrdering,
                                                     getCompilerContext().getMaximalPossibleTableCount(),
                                                     lcc,
-                                                    lcc.getJoinCostEstimationModel());
+                                                    lcc.getCostModel());
         }
 
         optimizer.prepForNextRound();
