+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/conn/GenericLanguageConnectionContext.java
@@ -50,8 +50,8 @@ import com.splicemachine.db.iapi.services.property.PropertyUtil;
 import com.splicemachine.db.iapi.services.sanity.SanityManager;
 import com.splicemachine.db.iapi.sql.*;
 import com.splicemachine.db.iapi.sql.compile.*;
-import com.splicemachine.db.iapi.sql.compile.costing.JoinCostEstimationModel;
-import com.splicemachine.db.iapi.sql.compile.costing.JoinCostEstimationModelRegistry;
+import com.splicemachine.db.iapi.sql.compile.costing.CostModel;
+import com.splicemachine.db.iapi.sql.compile.costing.CostModelRegistry;
 import com.splicemachine.db.iapi.sql.conn.*;
 import com.splicemachine.db.iapi.sql.depend.DependencyManager;
 import com.splicemachine.db.iapi.sql.depend.Provider;
@@ -519,9 +519,9 @@ public class GenericLanguageConnectionContext extends ContextImpl implements Lan
                     favorIndexPrefixIteration.equalsIgnoreCase("true")) {
                 this.sessionProperties.setProperty(SessionProperties.PROPERTYNAME.FAVORINDEXPREFIXITERATION, "TRUE".toString());
             }
-            String joinCostEstimationModel = connectionProperties.getProperty(Property.JOIN_COST_ESTIMATION_MODEL);
-            if(joinCostEstimationModel != null && JoinCostEstimationModelRegistry.exists(joinCostEstimationModel)) {
-                this.sessionProperties.setProperty(SessionProperties.PROPERTYNAME.JOINCOSTESTIMATIONMODEL, joinCostEstimationModel);
+            String costModel = connectionProperties.getProperty(Property.COST_MODEL);
+            if(costModel != null && CostModelRegistry.exists(costModel)) {
+                this.sessionProperties.setProperty(SessionProperties.PROPERTYNAME.COSTMODEL, costModel);
             }
         }
         if (type.isSessionHinted()) {
@@ -4134,12 +4134,12 @@ public class GenericLanguageConnectionContext extends ContextImpl implements Lan
     }
 
     @Override
-    public JoinCostEstimationModel getJoinCostEstimationModel() {
-        String joinCostEstimationModeName = getSessionProperties().getPropertyString(SessionProperties.PROPERTYNAME.JOINCOSTESTIMATIONMODEL);
-        if(JoinCostEstimationModelRegistry.exists(joinCostEstimationModeName)) {
-            return JoinCostEstimationModelRegistry.getJoinCostEstimationModel(joinCostEstimationModeName);
+    public CostModel getCostModel() {
+        String costModelName = getSessionProperties().getPropertyString(SessionProperties.PROPERTYNAME.COSTMODEL);
+        if(CostModelRegistry.exists(costModelName)) {
+            return CostModelRegistry.getCostModel(costModelName);
         } else {
-            return JoinCostEstimationModelRegistry.getJoinCostEstimationModel("v1");
+            return CostModelRegistry.getCostModel("v1");
         }
     }
