+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/conn/SessionPropertiesImpl.java
@@ -132,8 +132,8 @@ public class SessionPropertiesImpl implements SessionProperties {
                 boolean favorIndexPrefixIteration = Boolean.parseBoolean(valString);
                 properties[FAVORINDEXPREFIXITERATION.getId()] = favorIndexPrefixIteration;
                 break;
-            case JOINCOSTESTIMATIONMODEL:
-                properties[JOINCOSTESTIMATIONMODEL.getId()] = valString;
+            case COSTMODEL:
+                properties[COSTMODEL.getId()] = valString;
                 break;
             default:
                 assert false;
