+++ db-shared/src/main/java/com/splicemachine/db/iapi/reference/Property.java
@@ -1292,6 +1292,6 @@ public interface Property {
 
     String SPLICE_OLAP_PARALLEL_PARTITIONS = "splice.olapParallelPartitions";
 
-    String JOIN_COST_ESTIMATION_MODEL = "joinCostEstimationModel";
+    String COST_MODEL = "costModel";
 }
