+++ splice_machine/src/main/java/com/splicemachine/derby/impl/sql/compile/SpliceLevel2OptimizerFactoryImpl.java
@@ -25,7 +25,7 @@ import com.splicemachine.db.iapi.sql.compile.OptimizablePredicateList;
 import com.splicemachine.db.iapi.sql.compile.Optimizer;
 import com.splicemachine.db.iapi.sql.compile.OptimizerFactory;
 import com.splicemachine.db.iapi.sql.compile.RequiredRowOrdering;
-import com.splicemachine.db.iapi.sql.compile.costing.JoinCostEstimationModel;
+import com.splicemachine.db.iapi.sql.compile.costing.CostModel;
 import com.splicemachine.db.iapi.sql.conn.LanguageConnectionContext;
 import com.splicemachine.db.iapi.sql.dictionary.DataDictionary;
 import com.splicemachine.db.impl.sql.compile.OptimizerFactoryImpl;
@@ -55,7 +55,7 @@ public class SpliceLevel2OptimizerFactoryImpl extends OptimizerFactoryImpl {
                                   RequiredRowOrdering requiredRowOrdering,
                                   int numTablesInQuery,
                                   LanguageConnectionContext lcc,
-                                  JoinCostEstimationModel joinCostEstimationModel) throws StandardException {
+                                  CostModel costModel) throws StandardException {
         /* Get/set up the array of join strategies.
          * See comment in boot().  If joinStrategySet
          * is null, then we may do needless allocations
@@ -81,7 +81,7 @@ public class SpliceLevel2OptimizerFactoryImpl extends OptimizerFactoryImpl {
                                 requiredRowOrdering,
                                 numTablesInQuery,
                                 lcc,
-                                joinCostEstimationModel);
+                                costModel);
     }
 
 
@@ -92,7 +92,7 @@ public class SpliceLevel2OptimizerFactoryImpl extends OptimizerFactoryImpl {
             RequiredRowOrdering requiredRowOrdering,
             int numTablesInQuery,
             LanguageConnectionContext lcc,
-            JoinCostEstimationModel joinCostEstimationModel) throws StandardException {
+            CostModel costModel) throws StandardException {
 
         return new SpliceLevel2OptimizerImpl(
                 optimizableList,
@@ -107,7 +107,7 @@ public class SpliceLevel2OptimizerFactoryImpl extends OptimizerFactoryImpl {
                 requiredRowOrdering,
                 numTablesInQuery,
                 lcc,
-                joinCostEstimationModel);
+                costModel);
     }
 
     /**
